Split commands into seperate files
